Create new Sidebar component (#374)
This is a work in progress and features/fixes will be added to it when needed. For now, it supports a single level list of navigation items, a header and a responsive mode for mobile devices.

For desktops, the `Sidebar` should be always visible on the left side, and for mobile devices, it should be able to be visible (with a backdrop and a close button, as designed) or hidden.