Create Header component and SidebarContext to control toggling inside circuit (#375)
# Purpose
This PR aims to add the new `Header` component and the `SidebarContext` to control the toggling logic of the `Sidebar` inside circuit, so we can have similar usages on all applications.

# Side Notes
The `Header` is only visible on mobile devices, on desktops, it should hide and the `Sidebar` should always be open.

Also, the `Hamburger` was redesigned to have a border. The API was kept the same, besides the addition of a `light` prop for light or dark colors in different backgrounds.