Design documentation Pt. I (#359)
This is the first part of the documentation of the whole Circuit component set. The basic framework involves a definition of what is the component, more context to understand when to use it and overall usage guidelines with quick dos and dont's.

Included in this PR are forms, list, modal, spacing, and tooltip.