fix: Move minus symbol in front of the currency symbol
The minus symbol is put in front of the amount and currency symbol.
Before this change negative GBP and USD amounts were rendered
as £-1.71 and $-5.00 (basically every currency that appears before
the number amount) and after this change they are
rendered as -£1.71 and -$5.00.

Some examples using the same approach to negative currency amounts:

https://docs.microsoft.com/en-us/globalization/locale/currency-formatting
https://english.stackexchange.com/a/124804