feat(components): improve clear button on inputs (#559)
* feat(components): increase tap target size on input prefix and suffix

By switching from margin to padding for the spacing, the prefix and suffix elements get a much larger tap target size: 16px to 40px. This is especially important on touchscreen devices, but also improves usability for mouse users.

* feat(components): improve accessibility of the SearchInput clear button

* fix(components): deprecate showClear in favor of onClear prop

This brings the AutocompleteInput API in line with the SearchInput API. It also fixes the functionality of the clear button on the AutocompleteInput.