feat(components): rewrite the Selector as radio/checkbox
This makes the Selector component accessible (it wasn't before). It also brings the API inline with
the Checkbox and RadioButton components. Furthermore, it allows for multiple options to be selected.

BREAKING CHANGE:
The Selector component now accepts the onChange, value, name, and checked props. The onClick and
selected props have been removed.

ISSUES CLOSED: #544