feat(components): polish Button components (#605)
* feat(components): revert default Button variant to secondary

The previous default Button style matches the new secondary variant most, so to avoid a breaking change, I've made it the default variant.

* feat(components): use IconButton for Carousel buttons

* feat(components): change Hamburger default size to match IconButton

* feat(components): remove Button giga size

To our knowledge, this size was never used. In order to reduce the API surface, we're removing it.

* feat(components): tweak Button border radius

* test(components): update snapshots

* fix(components): fix small icons for Carousel buttons

BREAKING CHANGE:
The Button's `giga` size has been removed. Use the `mega` size (default) instead.