Migrate Table to TypeScript (#852)
* feat(components): migrate core Table component to TypeScript

* fix(tests): fix unused variable error and update snapshots

* feat(components): migrate SortArrow to TypeScript

* feat(components): migrate TableCell to TypeScript

* refactor: amend tests to use getByRole instead of testids

This also passes rowheaders={false} to the rendered component because it's not what's being testes here. With row headers, the first column of the table receives a "rowheader" WAI role instead of the "cell" role that we're targeting in the tests.

* fix: forward scope prop to table header cells

* feat(components): migrate TableRow to TypeScript

This commit also removes a custom keyboard event listener. Enter and space key down events are already handled as a click event. We were effectively triggering onClick twice - this was discovered by switching from fireEvent to userEvent. I'm leaving the test covering the keyboard event logic in the spec.

* fix(tests): update snapshots

* fix(components): mark data-testid prop in TableCell as optional

It will be assigned a value via defaultProps but should be optional from the point of view of its parents. This was breaking specs that were not explicitly passing a testid

* feat(components): migrate TableHeader to TypeScript

* feat(components): migrate TableHead to TypeScript

* feat(components): migrate TableBody to TypeScript

* fix(components) handle falsy sortByValue

If sortByValue evaluated to false, like  in the Stories, the ternary resolved to false and getSortByValue fell back to the cell's children instead of the sortByValue

* fix(components): pass a falsy sortedRow

If the sortedRow is 0, it evaluates to falsy, and we were passing undefined, breaking the UI of the SortArrow.

Note that the implementation will likely be changed when cleaning up the PR to make types match more across the Table components

* refactor: reuse types and switch defaults from null to undefined

* refactor: reuse Direction type across Table components

* fix(components): fix CellObject type to accept a ReactNode as children

* fix(components): reuse Cell type in Table component

* refactor: simplify types

* cleanup: remove leftover comments in utils.ts

* fix: address code review feedback

* fix: revert renaming of height attr and update types

* fix: reuse types from component props in static TableHeader methods

* fix: make TableProps extend HTML table element and spread props to wrapper

* Add changesets

* Specify that changeset applies to the Table component

Co-authored-by: Connor Bär <connor-baer@users.noreply.github.com>

Co-authored-by: Connor Bär <connor-baer@users.noreply.github.com>