Enforce sortLabel on sortable tables
This commit also fixes a TS error where sortLabel was not declared on the Cell type.

It does not differentiate between header and body cells (same as before), which means that users might think that sorting-related props like sortLabel are available on body cells, where they are not. Unfortunately, making the types stricter here will require a larger refactoring around the table utils which are currently shared by the header and body cells alike.