Enforce sortLabel for sortable Table columns
This commit also includes a refactoring around how the sort options are handled, because we want to prevent the column from sorting if a sortLabel is not set. The main change is a new getSortParams util that is called in TableHead before the params are passed to the Table Header.