Add Avatar and ImageInput components (#922)
* Create draft Avatar and ImageInput components

* Add SVG placeholders for object and identity variants

* Fix linting

* Update Avatar prop names

* Wrap Avatar in label in ImageInput instead of passing an as prop

* Add icons and logic for removing an image

* Clear file input via ref when the clear button is clicked

A file input is always uncontrolled so a ref is necessary here: https://reactjs.org/docs/uncontrolled-components.html\#the-file-input-tag

* Add @jsxRuntime pragma to enable usage with React 17

* Fix focus state UI bug where the outline appeared behind the ActionButton

* Add active state styles

* Update Avatar docs

This also renames the variant prop options to object/identity to match the design specs. Person/business was inaccurate because a business should actually use the identity variant with a business placeholder (in development).

* Switch Avatar default variant to object

This matches the ImageInput defaults and makes the two components more interchangeable

* Add onChange and onClear callbacks with loading and error states

* Add snapshot tests for Avatar

The Avatar is a purely visual component, so relying on snapshot testing makes sense here. It particularly ensures that the placeholder SVGs are set as expected wheen no image is provided.

* Fix types in ImageInput spec to pass CI

* Switch from brightness() to pseudo-elements and clean up styles

* Do not call onChange with a falsy file param

* Update ImageInput docs

* Add initial tests for ImageInput

Snapshots for main states (default, invalid, existing image) and functional test of an image upload via userEvent.upload

* Add business logic tests for ImageInput

* Rename ImageInput into AvatarInput

* Add test for the logic of clearing an uploaded avatar

* Properly export the components

* Add pointer-events:none to spinner to fix FF loading UI

* Address code review feedback

* Optimize SVGs

* Batched improvements (see commit description)

- the border-radius of the zetta size was increased to 12px, the value will be moves to design tokens and used in the next Circuit major version
- a bug on Chrome where uploading the same image twice would not fire onChange the second time was fixed by clearing the DOM input element on click
- the AvatarInput was adapted to accept any component as a visual element, the Avatar or something else. A new story "Custom Component" was added to document it. I will add tests for it and rename AvatarInput back to ImageInput in follow-up commits

* Rename AvatarInput back to ImageInput

Since the last commit, it accepts any visual element, although it is still optimized for the Avatar component. This will enable teams to use the ImageInput for use cases beyond an avatar upload.

* Move invalid border to inset

This is to ensure that the button still stands out well even with invalid styles (from a design feedback)

* Move add button out of label element

- This makes the component's implementation cleaner
- It makes the tests cleaner because there's only one matching label that we can query with getByLabelText
- This commit also adds missing prop descriptions

* Add snapshot test for rendering with a custom component

* Write docs

* Add changeset

* Address PR review comments

* Make the Avatar's alt prop required

The ImageInput also exposes it as part of the custom component's signature, but it defaults to  in that case.

Accessibility docs were also updated and an extra usage example with a custom component was added to the ImageInput stories.

* Update Avatar spec with required alt prop