Improve Table component API (#960)
* Clean up leftover fixme comment

* Change custom onSortBy signature

nextDirection is now optional (instad of null) and therefore moved to the third argument

* Enforce sortLabel on sortable tables

This commit also fixes a TS error where sortLabel was not declared on the Cell type.

It does not differentiate between header and body cells (same as before), which means that users might think that sorting-related props like sortLabel are available on body cells, where they are not. Unfortunately, making the types stricter here will require a larger refactoring around the table utils which are currently shared by the header and body cells alike.

* Remove unnecessary data-testid from the TableCell

* Enforce sortLabel for sortable Table columns

This commit also includes a refactoring around how the sort options are handled, because we want to prevent the column from sorting if a sortLabel is not set. The main change is a new getSortParams util that is called in TableHead before the params are passed to the Table Header.

* Update Table exports

We should only export the main Table component and relevant types

* Revert "Remove unnecessary data-testid from the TableCell"

This reverts commit a94f6a3b8576dff914cf4e5c1a6474d275fdfd61.

The testids are still set as defaults used in other Table components for testing, so it makes seense to keep it in the TableCell as well.

* Update stories imports

* Add more coverage to getSortParams

* Add changesets

* Update changeset

* Remove default testids from Table components

* Pass onClick down to the SortArrow to avoid a potential a11y regression

* Fix header height bug with word-wrap

* Add missing changesets