Switch to the new JSX transform (#991)
* Remove jsxRuntime pragmas

* Migrate to the  jsx automatic runtime - part 1

* Migrate to the jsx automatic runtime - part 2

* Fix TS errors by opting into the new jsx runtime in the root tsconfig

* Fix Storybook List bug

* Fix React not defined errors with shorthand fragments

This can be fixed by either keeping the React import or importing Fragment and using it directly. This commit applies the second option in Circuit components

* Use babel-jest instead of ts-jest

* Update snapshots

* Add types for @emotion/core/jsx-runtime

The types were added in Emotion 11, we can remove this when we upgrade. Until then, we're typing the module globally to silence TS and hopefully fix the build errors