React 18 (#1554)
* Upgrade React deps

This updates the project's deps to React 18 and updates the peer deps range to add support for it

* Address React 18 breaking changes

Only type-related (for now)

* Skip lib checks again

* Resolve to React v17 types

Third-party packages (e.g. react-modal, react-number-format) depend on @types/react@*, which resolves to v18 by default, and throws type errors in components using the packages

* Upgrade storybook

* Upgrade testing lib packages

* Fix noMaring in SearchInput stories

* Fix SearchInput

We need to return a truthy suffix even when the close button is not rendered, so that the Input component applies padding consistently with and without the suffix. Otherwise, the Input's width will change depending on whether the user has typed something.

* Update Popover snapshots

* Fix false positive in ToastCoontext spec

* Fix failing ToastCoontext test

Unsure why it was working with the old versions of React/testing library

* Fix SidePanelContext specs

It looks like there is an issue with concurrency where if actions are grouped in a single act block, the first action doesn't resolve before the second nd the test fails (race condition). Maybe?

* Fix DesktopSidePanel spec

* Fix ModalContext specs

* Bump react-modal to latest

* Fix ts errors in CurrencyInput spec

* Bump number-fomat and fix spec

* Unskip lib checks in tsconfigs

* Remove unnecessary type annotations from styled components

* Add changeset