Replace renderToStaticMarkup test util by the container from render()
`renderToStaticMarkup` was being used in a11y tests because jest-axe expects either a string or an HTML element to run against.

However, it seems like it's using NodeJS's TextEncoder under the hood, and since we're running tests in jsdom, TextEncoder is not available.

An alternative solution would have been to inject the TextEncoder util as a global in the jsdom environment, but I couldn't get this to work on my first try, and it was recommended against by a jest maintainer, cf. https://github.com/jsdom/jsdom/issues/2524#issuecomment-902027138.

Instead, I followed the example on the jest-axe README and ran axe against the container returned by @testing-library's render method, itself an HTMLElement:

```
const { container } = render(<Component />);
```