Add locale to defaultProps
Since this is how the input is most commonly used, it makes sense to set the locale by default in tests.

This commit also adds a snapshot for the true default, when neither a locale nor a currency are passed (despite the TS error). In this case, the component falls back to DEFAULT_FORMAT (Â$ prefix).